Pasado de float a double en velocity_verlet y declaracion en .h

+++ src/avanzar.c
@@ -8,7 +8,7 @@
 
 // Definicion de funciones
 
-float velocity_verlet(double* x, double* v, double* f, double h, double m)
+double velocity_verlet(double* x, double* v, double* f, double h, double m)
 {
 	int i;
 	for(i = 0; i<3*N; i++)
